Get ISO3 code of all World countries

Get the ISO3 code of all World countries from the Protected Planet website (based on the UNEP classification).

get_isocodes()

Value

A data frame with 3 columns: - region, the regions of the world - country, the countries of the world - iso3, the iso3 code of countries

Details

A Internet connexion is required.

Examples

x <- get_isocodes()

head(x)

#>   region      country iso3
#> 1 Africa      Algeria  DZA
#> 2 Africa      Burundi  BDI
#> 3 Africa       Angola  AGO
#> 4 Africa     Botswana  BWA
#> 5 Africa        Benin  BEN
#> 6 Africa Burkina Faso  BFA

x[x[ , "region"] == "Africa", ]

#>    region                                      country iso3
#> 1  Africa                                      Algeria  DZA
#> 2  Africa                                      Burundi  BDI
#> 3  Africa                                       Angola  AGO
#> 4  Africa                                     Botswana  BWA
#> 5  Africa                                        Benin  BEN
#> 6  Africa                                 Burkina Faso  BFA
#> 7  Africa                                        Egypt  EGY
#> 8  Africa                     Central African Republic  CAF
#> 9  Africa                                     Cameroon  CMR
#> 10 Africa                 Democratic Republic of Congo  COD
#> 11 Africa                                        Congo  COG
#> 12 Africa                                        Gabon  GAB
#> 13 Africa                                Côte D'Ivoire  CIV
#> 14 Africa                                        Ghana  GHA
#> 15 Africa                                       Guinea  GIN
#> 16 Africa                                       Gambia  GMB
#> 17 Africa                                Guinea-Bissau  GNB
#> 18 Africa                                        Libya  LBY
#> 19 Africa                                       Malawi  MWI
#> 20 Africa                                      Morocco  MAR
#> 21 Africa                                        Kenya  KEN
#> 22 Africa                                      Lesotho  LSO
#> 23 Africa                                   Madagascar  MDG
#> 24 Africa                                   Mozambique  MOZ
#> 25 Africa                                    Mauritius  MUS
#> 26 Africa                                       Rwanda  RWA
#> 27 Africa                                        Niger  NER
#> 28 Africa                                      Nigeria  NGA
#> 29 Africa                                      Tunisia  TUN
#> 30 Africa                        São Tomé and Príncipe  STP
#> 31 Africa                                        Sudan  SDN
#> 32 Africa                               Western Sahara  ESH
#> 33 Africa Saint Helena, Ascension and Tristan da Cunha  SHN
#> 34 Africa                                       Uganda  UGA
#> 35 Africa                                       Zambia  ZMB
#> 36 Africa                                     Zimbabwe  ZWE
#> 37 Africa                                      Mayotte  MYT
#> 38 Africa                                      Réunion  REU
#> 39 Africa                            Equatorial Guinea  GNQ
#> 40 Africa                                         Chad  TCD
#> 41 Africa                                     Djibouti  DJI
#> 42 Africa                                      Eritrea  ERI
#> 43 Africa                                     Ethiopia  ETH
#> 44 Africa                                      Somalia  SOM
#> 45 Africa                                  South Sudan  SSD
#> 46 Africa                  United Republic of Tanzania  TZA
#> 47 Africa                                      Comoros  COM
#> 48 Africa                                      Namibia  NAM
#> 49 Africa                     Eswatini, the Kingdom of  SWZ
#> 50 Africa                                   Seychelles  SYC
#> 51 Africa                                 South Africa  ZAF
#> 52 Africa                                   Cabo Verde  CPV
#> 53 Africa                                      Liberia  LBR
#> 54 Africa                                         Mali  MLI
#> 55 Africa                                   Mauritania  MRT
#> 56 Africa                                      Senegal  SEN
#> 57 Africa                                 Sierra Leone  SLE
#> 58 Africa                                         Togo  TGO