Download Protected Areas by Country

This function downloads protected areas for one country using the WDPA API.

get_wdpa(isocode, key = "WDPA_KEY")

Arguments

isocode
a character specifying the ISO-3 code of the country.

key
a character providing the WDPA token name stored in the .Renviron file. Default is WDPA_KEY.

Value

A MULTIPOLYGON Simple Feature of protected areas defined in the EPSG 4326 with the following 19 attributes:

wdpa_id

the ID of the protected area on the WDPA database

pa_name

the name of the protected area on the WDPA database

original_name

the original name of the protected area

country_iso3

the ISO-3 code of the country

country

the name of the country

owner_type

the owner of the protected area

is_marine

a boolean: TRUE for marine and FALSE for terrestrial protected area

designation

the designation of the protected area

iucn_category

the IUCN category

no_take_status

the no take status

reported_area

the (reported) area of the protected area

reported_marine_area

the (reported) area of the marine part of the protected area

no_take_area

the (reported) area of the no take part of the protected area

legal_status

the legal status of the protected area

management_authority

the management authority

governance

the type of gouvernance

management_plan

the URL of the management plan

protected_planet

the URL of the protected area page on the Protected Planet website

legal_status_updated

the date of the last update of the legal status

Spatial shapefiles (ESRI format) are also written on the hard drive in a new folder XXX_protectedareas/ (where XXX is the ISO-3 code of the country).

See also

Examples

if (FALSE) {
## Get the ISO-3 code of Georgia ----
data("wdpa_countries", package = "worldpa")
wdpa_countries[grep("^Georgia", wdpa_countries$"country_name"), 3:4]

##                                     country_name country_iso3
## 133                                      Georgia          GEO
## 235 South Georgia and the South Sandwich Islands          SGS

## Download Protected Areas for Georgia ----
worldpa::get_wdpa(isocode = "GEO")
}